Move Katib Charms to canonical org
Given that we now have an automated way of publishing the Charmed
Kubeflow charms we will need to have a repository with the katib charms
in our own org to benefit from these automations.

The Katib Charms will still exist in the upstream kubeflow org and we
will sync the charms with upstream during new releases.

This commit moves the charms to a similar structure as our
other multi charm repositories. It also adds the release and publish
automations to the github actions. The integration tests remain the same
and will have to be modernized at another point.